+++ packages/thirdweb/src/exports/react.ts
@@ -97,5 +97,8 @@ export {
   type AutoConnectProps,
 } from "../react/core/hooks/connection/useAutoConnect.js";
 
+// auth
+export { type SiweAuthOptions } from "../react/core/hooks/auth/useSiweAuth.js";
+
 export { useBuyWithFiatQuote } from "../react/core/hooks/pay/useBuyWithFiatQuote.js";
 export { useBuyWithFiatStatus } from "../react/core/hooks/pay/useBuyWithFiatStatus.js";