+++ packages/thirdweb/src/extensions/marketplace/english-auctions/write/executeSale.ts
@@ -35,7 +35,6 @@ export function executeSale(
         contract: options.contract,
         auctionId: options.auctionId,
       });
-      console.log("*** winning bid", winningBid);
       if (!winningBid) {
         throw new Error("Auction is still active");
       }